requirements.txt
PyQt6>=6.5.0
cryptography>=41.0.0
psutil>=5.9.0
pyyaml>=6.0